refactor: add ApplicationConfig for application-wide settings
