+++ src/main/kotlin/com/dobby/backend/config/ApplicationConfig.kt
@@ -1,0 +1,10 @@
+package com.dobby.backend.config
+
+import com.dobby.backend.infrastructure.config.TokenProperties
+import org.springframework.boot.context.properties.EnableConfigurationProperties
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableConfigurationProperties(TokenProperties::class)
+class ApplicationConfig {
+}
